backend/pyproject.toml
[project]
name = "marketing-ai-agency"
version = "0.1.0"
description = "AI-powered marketing agency with Google ADK agents"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.0.0",
    "google-genai>=1.0.0",
    "google-cloud-aiplatform>=1.74.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "websockets>=13.0",
    "pydantic>=2.0.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
